refactor(oracle): migrate dns generator sampler and behavior
